Moved settings.tsx into pages folder
